refactor: implement Null pattern classes for services and add service installation rollback logic
